parsegrammarfile.py
Made a new module parsegrammarfile.py. It takes the list of lists from readgrammarfile.py and parses it to return a grammar class which will be used later on to generate the sentence fragment. It also goes through each list in the list of lists (each is a rule), and stores its properties in their corresponding recursive objects. Also created a unittest for the module.