variablesymbol.py + grammar.py
Created a new sub function in the Grammar class, get_rule() which gets the rule associated with given variable. Also finished the generate_sentence_fragment() function for the variable class. However, for the uniitest I am unable to currently test as I do not have the Rule class implemented to use to get the variable associated with the symbol.